backend/requirements.txt
# ============================================
# Knowledge Assistant - Dependencies
# Updated: 2025-12-22 (Latest Stable Versions)
# ============================================

# FastAPI & Web Framework
fastapi>=0.127.0,<0.128.0          # Latest: 0.127.0
uvicorn[standard]>=0.40.0          # Latest: 0.40.0 - ASGI server
python-multipart>=0.0.18           # File uploads
sse-starlette>=2.2.0               # Server-Sent Events
pydantic>=2.12.0,<3.0.0            # Latest: 2.12.5
pydantic-settings>=2.7.0           # Settings management

# LLM Integration
litellm>=1.80.0,<2.0.0             # Latest: 1.80.10 - Unified LLM interface
openai>=1.60.0                     # OpenAI API client

# Note: LangChain/LangGraph/llama-index removed - not used in current codebase
# Uncomment below if needed for future agent development:
# langchain>=0.3.0
# langchain-community>=0.3.0
# langgraph>=0.2.0

# Vector Store & Embeddings
chromadb>=1.3.0,<2.0.0             # Latest: 1.3.7 - Vector database
sentence-transformers>=3.3.0       # Latest: 3.3.1 - Embeddings
FlagEmbedding>=1.2.11              # Latest: 1.2.11 - BGE-M3 support

# Document Processing
python-docx>=1.1.0                 # DOCX parsing
pypdf>=5.1.0                       # Latest: 5.1.0 - PDF parsing
markdown>=3.7.0                    # Markdown processing
mistune>=3.0.0                     # Fast Markdown parser
beautifulsoup4>=4.12.0             # HTML parsing

# Database & ORM
sqlalchemy>=2.0.45,<3.0.0          # Latest: 2.0.45 - Async support
aiosqlite>=0.20.0                  # SQLite async driver
alembic>=1.14.0                    # Database migrations

# Task Queue (Optional)
arq>=0.26.0                        # Latest: 0.26.3 - Async task queue
redis>=5.2.0                       # Redis client

# Authentication & Security
PyJWT>=2.10.0                      # JWT handling (replaced python-jose)
pwdlib[argon2,bcrypt]>=0.2.0       # Password hashing (replaced passlib)

# HTTP & Networking
httpx>=0.28.0                      # Latest: 0.28.1 - Async HTTP client
aiofiles>=24.1.0                   # Async file I/O

# Configuration & Logging
python-dotenv>=1.0.1               # Environment variables
structlog>=24.4.0                  # Structured logging
tenacity>=9.0.0                    # Retry logic

# Observability & Monitoring
opentelemetry-api>=1.29.0          # OpenTelemetry API
opentelemetry-sdk>=1.29.0          # OpenTelemetry SDK
opentelemetry-instrumentation-fastapi>=0.50b0  # FastAPI instrumentation

# Testing
pytest>=8.3.0                      # Testing framework
pytest-asyncio>=0.24.0             # Async test support
pytest-cov>=6.0.0                  # Coverage reporting

# Development & Code Quality
ruff>=0.8.0                        # Fast linter/formatter
mypy>=1.13.0                       # Type checking
pre-commit>=4.0.0                  # Git hooks
